Make TF Lite compile after rebase
The merge CL that is a parent of this one was done by accepting all
incoming changes and addressing build errors afterwards. Here is what
was done to make TF Lite and NNAPI compile after this update:
* bfloat16 implementation was moved to eigen upstream. I've restored it
in the original place, so that we don't have to upgrade eigen to an
unstable version yet.
* Disabled log sinks, reverted to old implementation of
LogMessage::GenerateLogMessage()
* Added new source files to Android.bp files
* Switched code to use farmhash from libtextclassifier
* Switched absl::optional to std::optional
* Added a genrule for lite/schema/mutable/schema_generated.h
* Ruy was moved to its own project external/ruy
* Temporary disabled registration of RFFT2D op until the corresponding
fft library is imported to Android

Bug: 178609672
Test: mma, NeuralNetworksTest_static
Change-Id: I97088e763bf430390e5becfee5bb937088613b18