Remove most local modifications.
The previous merges from upstream in aosp/2193490 and aosp/1793621
introduced a lot of modifications that seem to be merge errors.
This makes further updating the library very difficult. Revert most
of them and only leave modifications that were made for specific
functional reasons.

This CL also corrects the METADATA entry so that it accurately
records the last merged upstream commit.

Bug: 302418620
Test: mm, CTS tests
Flag: EXEMPT third-party library update
Change-Id: I31b7af4ec1cd520ce1e892064324cd2fd7f21993